refactor: add shared pagination component
