+++ view/components/Shared/Feed.tsx
@@ -4,15 +4,14 @@ import {
   Card,
   CardContent,
   SxProps,
-  TablePagination,
   Typography,
 } from '@mui/material';
-import { ChangeEvent, MouseEvent, useState } from 'react';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FeedItemFragment } from '../../graphql/posts/fragments/gen/FeedItem.gen';
-import { useIsDesktop } from '../../hooks/shared.hooks';
 import PostCard from '../Posts/PostCard';
 import ProposalCard from '../Proposals/ProposalCard';
+import Pagination from './Pagination';
 
 const CARD_CONTENT_STYLES: SxProps = {
   '&:last-child': {
@@ -39,19 +38,6 @@ const Feed = ({ feed, ...boxProps }: Props) => {
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
   const { t } = useTranslation();
-  const isDesktop = useIsDesktop();
-
-  const handleChangePage = (
-    _: MouseEvent<HTMLButtonElement> | null,
-    newPage: number,
-  ) => setPage(newPage);
-
-  const handleChangeRowsPerPage = (
-    event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
 
   if (feed.length === 0) {
     return (
@@ -67,35 +53,22 @@ const Feed = ({ feed, ...boxProps }: Props) => {
 
   return (
     <Box {...boxProps}>
-      <TablePagination
-        component="div"
-        count={100}
+      <Pagination
         page={page}
-        onPageChange={handleChangePage}
         rowsPerPage={rowsPerPage}
-        onRowsPerPageChange={handleChangeRowsPerPage}
-        SelectProps={{
-          sx: { display: isDesktop ? undefined : 'none' },
-        }}
-        labelRowsPerPage={isDesktop ? undefined : <></>}
-        sx={{ marginBottom: 1.5 }}
+        setPage={setPage}
+        setRowsPerPage={setRowsPerPage}
       />
 
       {feed.map((item) => (
         <FeedItem item={item} key={`${item.__typename}-${item.id}`} />
       ))}
 
-      <TablePagination
-        component="div"
-        count={100}
+      <Pagination
         page={page}
-        onPageChange={handleChangePage}
         rowsPerPage={rowsPerPage}
-        onRowsPerPageChange={handleChangeRowsPerPage}
-        SelectProps={{
-          sx: { display: isDesktop ? undefined : 'none' },
-        }}
-        labelRowsPerPage={isDesktop ? undefined : <></>}
+        setPage={setPage}
+        setRowsPerPage={setRowsPerPage}
       />
     </Box>
   );
